[2607-DEV-496] Step 3: extract ABO verify/approve orchestration from route into lib/abo/verifyAbo.ts
